Feat: 학과 선택 페이지에 서스펜스 적용
